+++ packages/core/src/lib/server/createExpressServer.ts
@@ -6,8 +6,8 @@ import { graphqlUploadExpress } from 'graphql-upload';
 import { ApolloServer } from 'apollo-server-express';
 import type { KeystoneConfig, CreateContext, SessionStrategy, GraphQLConfig } from '../../types';
 import { createSessionContext } from '../../session';
-import { DEFAULT_FILES_STORAGE_PATH } from '../context/createFilesContext';
-import { DEFAULT_IMAGES_STORAGE_PATH } from '../context/createImagesContext';
+import { DEFAULT_FILES_STORAGE_PATH } from '../assets/createFilesContext';
+import { DEFAULT_IMAGES_STORAGE_PATH } from '../assets/createImagesContext';
 import { createApolloServerExpress } from './createApolloServer';
 import { addHealthCheck } from './addHealthCheck';
